career-mentor-agent/requirements.txt
chainlit
python-dotenv
openai
openai-agent-sdk
requests 